fix: package dao, package mapper, package service, package test

+++ src/main/java/ru/yandex/practicum/filmorate/service/user/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service.user.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.*;
 import ru.yandex.practicum.filmorate.model.User;
@@ -12,20 +13,18 @@ import java.util.Collection;
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Primary
 public class UserServiceImpl implements UserService {
     private final UserStorage userStorage;
 
     @Override
     public User getUserById(Long id) {
-        checkUserExistence(id);
-        return userStorage.getUserById(id);
-    }
-
-    private void checkUserExistence(Long id) {
-        if (userStorage.getUserById(id) == null) {
+        User user = userStorage.getUserById(id);
+        if (user == null) {
             log.warn("User with id={} not found", id);
             throw new UserNotFoundException(String.format("User with id=%d not found", id));
         }
+        return user;
     }
 
     @Override
@@ -36,15 +35,23 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Collection<User> getFriends(Long id) {
-        getUserById(id);
+        isExistsIdUser(id);
         log.info("Received a list of friends for user id={}", id);
         return userStorage.getFriends(id);
     }
 
+    private void isExistsIdUser(Long userId) {
+        boolean isExists = userStorage.isExistsIdUser(userId);
+        if (!isExists) {
+            log.warn("User with id={} not found", userId);
+            throw new UserNotFoundException(String.format("User with id=%d not found", userId));
+        }
+    }
+
     @Override
     public Collection<User> getMutualFriends(Long id, Long otherId) {
-        getUserById(id);
-        getUserById(otherId);
+        isExistsIdUser(id);
+        isExistsIdUser(otherId);
         log.info("Received a list of mutual friends of user id={} and user otherId={}", id, otherId);
         return userStorage.getMutualFriends(id, otherId);
     }
@@ -52,9 +59,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public User createUser(User user) {
         setUserNameIfMissing(user);
-        User createdUser = userStorage.createUser(user);
-        log.info("Create user {}", user);
-        return createdUser;
+        User userCreated = userStorage.createUser(user);
+        log.info("Create user {}", userCreated);
+        return userCreated;
     }
 
     private void setUserNameIfMissing(User user) {
@@ -66,36 +73,40 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateUser(User user) {
-        checkUserExistence(user.getId());
+        isExistsIdUser(user.getId());
         setUserNameIfMissing(user);
-        User updatedUser = userStorage.updateUser(user);
         log.info("Update user {}", user);
-        return updatedUser;
+        return userStorage.updateUser(user);
     }
 
     @Override
     public User addInFriend(Long id, Long friendId) {
-        checkUserExistence(id);
-        checkUserExistence(friendId);
-        log.info("User id={} and user id={} are added as friends to each other", id, friendId);
-        return userStorage.addInFriend(id, friendId);
+        isExistsIdUser(id);
+        isExistsIdUser(friendId);
+        User user = userStorage.addInFriend(id, friendId);
+        log.info("User id={} has been added as a friend to user id={} with status={}", friendId, id, true);
+        return user;
     }
 
     @Override
     public String deleteFromFriends(Long id, Long friendId) {
-        checkUserExistence(id);
-        checkUserExistence(friendId);
-        userStorage.isExistsFriendship(id, friendId);
+        isExistsIdUser(id);
+        isExistsIdUser(friendId);
+        boolean isExistsFriendship = userStorage.isExistsFriendship(id, friendId);
+        if (!isExistsFriendship) {
+            log.info("Friendship user id={} with user id={} not found", friendId, id);
+            return String.format("Friendship user id=%d with user id=%d not found", friendId, id);
+        }
         userStorage.deleteFromFriends(id, friendId);
-        log.info("User id={} and user id={} are deleted as friends to each other", id, friendId);
-        return String.format("User id=%d and user friendId=%d are deleted as friends to each other", id, friendId);
+        log.info("User id={} has been removed from friends of user id={}", friendId, id);
+        return String.format("User id=%d has been removed from friends of user id=%d", friendId, id);
     }
 
     @Override
     public String deleteUser(Long id) {
-        checkUserExistence(id);
+        isExistsIdUser(id);
         userStorage.deleteUser(id);
-        log.info("User id={} removed", id);
-        return String.format("User id=%d deleted", id);
+        log.info("User with id={} deleted", id);
+        return String.format("User with id=%d deleted", id);
     }
 }
